pipeline: support for register renaming in dispatch stage
Fix Description
===============
    * New routine implementing the register renaming process.
    * Many util routines to supplement register renaming.
    * New name is generated by a non-decreasing 32-bit counter. No need to
      worry about integer overflow: 10k instrctions and 3 regs per inst = 30k
      inst, << 2^32 values.
    * Pretty print code to dump whole RMT or just an entry of it.
    * TODO: Write issue handler.
    * TODO: Fix freeing the inst while deleting from dispatch list after moving
      the inst to issue list.

Stats
=====
 src/dis-pipeline-pri.h |  41 +++++++++++++++++++++++--
 src/dis-pipeline.c     | 119 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++--------
 src/dis-pipeline.h     |  37 +++++++++++++++++++++++
 src/dis-print.c        |  78 +++++++++++++++++++++++++++++++++++------------
 src/dis-print.h        |   4 +++
 src/dis-utils.h        |  10 +++---
 src/dis.c              |  21 +++++++++++--
 src/dis.h              |  21 ++++++++-----
 8 files changed, 282 insertions(+), 49 deletions(-)

+++ src/dis.h
@@ -87,9 +87,9 @@ struct dis_inst_data {
     uint8_t     state;              /* fetch/decode/dispatch... */
     uint32_t    pc;                 /* pc as given in trace     */
     uint8_t     type;               /* inst type - 0, 1, 2      */
-    uint8_t     dreg;               /* dst register             */
-    uint8_t     sreg1;              /* src register 1           */
-    uint8_t     sreg2;              /* src register 2           */
+    uint16_t    dreg;               /* dst register             */
+    uint16_t    sreg1;              /* src register 1           */
+    uint16_t    sreg2;              /* src register 2           */
     uint32_t    mem_addr;           /* mem address in trace     */
     uint32_t    cycle[STATE_MAX];   /* state-cycle transition   */
